refactor: Move support email fetching function to `logger` domain
